1000x1000 bedrock works
Jumping now considers the velocity as well as the current displacement
which fixes many edge cases. Before the bot would likely fail if jumping
in a three block pattern where the path between them has a 90 degree edge.
This is because the bot's heading only focused on the next block---it
did not have a factor to counter the velocity going _away_ from the
current target.